requirements.txt
yfinance
python-telegram-bot
google-generativeai